dist update
Added in support using for a distance prior. There is a hard-coded dependence
on `(l, b)` right now, but that should be fine given we probably will be
using this with respect to the Green et al. (2014) galaxy model and the
Bailer-Jones et al. (2018) exponential-volume prior.

Along with this, I've also modified the plotting utilities to incorporate
the distance prior.

As part of this update, the priors have been moved to
their own module (`pdf`). This cleans things up a bit code-wise within
`fitting`.