fit refactor
This update adds several improvements to the code:

- Fitting is now done in two stages: (1) fit assuming gaussian mag errors, and
(2) improve upon the best-fit solution until the fits converge. While (1) isn't
perfect, it gets pretty close so that only a handful of iterations are often
required to improve the fit. This drastically cuts down on the runtime, which
is now ~2.2us/model/band/object! For the default MIST grid, this translates to
~0.9s/band/object.

- Fitting has been refactored so that the optimization step happens in an
internal function. This makes it easier to modify that particular portion of
the algorithm.

This now should pave the way to add in the Bayestar grid.