v0.2.0
This is a pretty significant update to the code. It makes enough
changes that I think the code should be relatively stable going forward,
and I'm bringing it out of "alpha" into "beta" development as a result.

Big changes:
- Removed dependencies on `minesweeper` when installing. Running
  `seds` still requires the relevant files to be installed.
- Distance and reddening draws are now (by default) saved during runtime.
  This increases storage space by ~20% but significantly reduces
  the amount of post-processing needed.
- Filters now have been moved to `filters.py`.
- Split up functions into a new `utils` modules which I've started
  fleshing out.

Bugfixes:
- The PS1 absolute r-band magnitude luminosity function prior should
  now properly install.
- Fixed a bug when smoothing which was causing the relevant smoothing in the
  x-direction to be systematically off in a non-trivial way. This now
  happens during post-processing and also includes parallax-based checks.
- Fixed some logic bugs when running plotting/utility functions.

I also moved the demos to a dedicated folder and have updated them
accordingly. I'm still in the process of finishing the bayestar comparison,
so that one will be updated shortly.