sample-based LOS
Implemented during the Finkbeiner Group's 2018 "Cape Code"
trip. This commit introduces a variant of the original LOS
fit. Originally, the LOS used binned PDFs, with the LOS
evaluated over the bins (with possibly some linear
interpolation). Here, I introduce an alternate approach using
the samples stored from the initial fits. Essentially, we weight
(a subset of) samples along the LOS according to a specified
kernel with some smoothing `s`. This is covariant with our
outlier model `pb` and allows for adaptive smoothing based on the
scatter along the LOS.

To summarize:
- Added new samples-based likelihood for the LOS fit.
- Added Lorentzian, Gaussian, and Tophat kernels for the LOs fit.
- Changed some details about the way the likelihood/prior
transform is evaluated.
- Suppressed some common (unhelpful) errors
- Added in some additional tags to `setup.py`.