v0.4.0
This commit finalizes up things for v0.4.0. The largest update
is the introduction of a sample-based LOS fit with adaptive
smoothing. This particular commit includes:
- small changes to the `prior_transform` API to go between the
binned and sample-based versions using the new `smooth` arg
- the introduction of a new foreground smoothing argument `s0`
so that the foreground smoothing isn't tied to the behavior of
the cloud
- more detailed comparisons with bayestar on
a different sightline (Orion_Lam)