v0.4.2
This commit introduces some changes to computing fits with `loglikelihood`
by allowing models to be "culled" before the optimization step using
their initial magnitude-based log-likelihoods. While this introduces a small
amount of additional overhead, in general it improves the overall
runtime on a number of testbeds by ~20-30%.

This commit also changes the titles to show the 95% credible
intervals to match the default quantiles, as well as the default
opacity of the priors (so that the posteriors can be more easily seen).