filter masks
Changed some of the internal methods by which the binary
filter masks were being implemented.
The way I previously had things was a holdover
from `frankenz`, where you could have models missing
data and so need to actually do the pairwise
check for every band. In `brutus`, we'll only be fitting
where our models are defined, and so
there's no need to check all the bands when we can
just sub-select the ones that are relevant.
This adds some memory overhead since we now are
creating new model arrays internally each time, but in
general this does prevent issues where runtime scales as
~total number of bands, rather than ~actual number of bands.
This appears to grant the expected ~Nbands_obs/Nbands_tot
speedup, although on machines with slower memory allocation
the overhead involved might be more of an issue.

Also modified `filters` description so it matches.