par->scale prior
This commit introduces the use of the parallax prior directly
in scale-factor space. This works generally when the parallax
has SNR > 4; below this, the default behavior is to avoid applying
it as a prior at all. This, combined with the removal of the
distance prior upon pre-selection, means I am much less aggressive
about pre-selecting models (based on the MLE scale-factors)
before subsequently integrating over the priors.
This helps some issues I was seeing where too many
models were getting killed early on.

Note:
I experimented with allowed the parallax to
also enter into the MLE scalefactor and Av, but eventually decided
the vanilla implementation should remain in place. I might come
back to this in the future.

Some other updates:
- Changed a number of defaults. Most of these involve increasing
various hyper-parameters, making runtimes longer but posteriors
more stable.
- Fixed a bug where the first iteration of `_optimize_fit`, based
on magnitudes, wasn't performing as expected, leading to a lot of
models being killed off early on when computing the likelihood.