fitting optimization
Adjusted the fitting subroutines to be a bit more optimized.
Allowed the parallax to be used as a prior to pre-emptively remove models
whose 5-sigma distance ranges have no overlap
with a provided parallax's 5-sigma distance range. This minimally affects
the fit but substantially improves the runtime to objects with high-SNR
parallaxes.

I might be able to squeeze out a few more speedups by trying
to adaptively cull the grid, but ultimately I think speedups targeting
huge grids are not feasible: the future direction of this work will lie
in figuring out ways to reduce the grid size so that MIST can look more
like Bayestar.

Fixed link to the v7 grid.