gaia MAW + tweaks
Updated the Gaia filter curves from DR2Rev to MAW. For BP, I've only added
in the faint filter curve for now just because I doubt we'll ever need
the bright curve (but other people might in the future). Models have been
updated, as have the neural network interpolate files.

Also slightly improved the `verbose` output of `BruteForce` so that
the user can see how the fits are doing in real-time, which is invaluable
when debugging.