fitting reorganization
Refactored/reorganized the internals within the `fitting` module.
This leads to changes to make `loglike` and `lnpost` more directly
accessible (and marks them as such) so users can experiment with
running those functions on particular objects directly. Also
should make future testing easier.

In addition, I've now added in a `_setup` function with
`BruteForce` that runs all of the relevant safety checks and
pre-processes all the data. This is run within `fit` and `_fit`.
This makes the internals cleaner and hopefully will be something
we can just modify directly if we want to do more things with
the data/models before running over each object.

Finally, this commit removes support for parallelization with
pools. It was clunky and ineffective, so since we are hoping to
do this with `numba` anyways I just removed it.