demos update
A preliminary update to the demos as I prepare to write up documentation
(since I like to do notebooks first, then structure the docs around
the notebooks). This now includes:

1. Models/priors, which goes over `SEDmaker`, `Isochrone`, and `pdf`.
2. Model grids, which outlines use of `SEDmaker.make_grid()`.
3. Fitting individual sources, which goes over `BruteForce`, interpreting
   output results, `plotting`, and relevant convenience
   functions in `utils`.

These are all updated so that they will work "out of the box" as of
`v0.8.2` (finally).

TODO: Add in demos for line-of-sight dust fitting, cluster modeling,
      and photometric offset modeling.

Also, I updated the default smoothing in `plotting.dist_vs_red` so that
it uses `0.015` rather than `0.01`, just since I've found that tends
to make slightly better-looking plots.