file download utilities
Resolves #44 by adding in some utility functions to download files,
along with a notebook demonstrating usage. This also coincides
with the creation of a Harvard Dataverse to store and archive
input files for posterity. README has been updated accordingly.

Note that `offsets_mist_v9` doesn't yet exist. I plan to derive those
ASAP and will update both the alias (which is currently broken in the code)
and the dataverse with those files as soon as I have them.