men/runtime improvements
This commit adds the `mem_lim` argument to `BruteForce.fit()`, which
allows users to limit the overall memory usage of the fitter
during runtime (ignoring overheads). This is only an approximate
upper limit but appears to work well during tests.

In addition, it also rolls back the `chi2min` change from the previous
commit to now be pre-resampling again (decreasing the typical values).
It also includes changes to the regularization to make it much
more aggressive after the first few iterations.